backend/requirements.txt
Django>=4.2
psycopg2-binary>=2.9
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
Pillow>=10.0
dj-database-url>=1.0
